+++ include/cloudbreak.bash
@@ -41,19 +41,21 @@ cloudbreak-conf-tags() {
     env-import CB_DOCKER_CONTAINER_AMBARI_WARM ""
 }
 
+consul-recursors() {
+    declare desc="Generates consul agent recursor option, by reading the hosts resolv.conf"
+    declare resolvConf=${1:? 'required 1.param: resolv.conf file'}
+    declare bridge=${2:? 'required 2.param: bridge ip'}
+
+    local nameservers=$(sed -n "/nameserver/ s/^.*nameserver[^0-9]*//p;" $resolvConf)
+    debug "nameservers on host:\n$nameservers"
+    debug bridge=$bridge
+    echo "$nameservers" | grep -v $bridge | sed -n '{s/^/ -recursor /;H;}; $ {x;s/[\n\r]//g;p}'
+}
+
 cloudbreak-conf-consul() {
     env-import DOCKER_CONSUL_OPTIONS ""
     if ! [[ $DOCKER_CONSUL_OPTIONS =~ .*recursor.* ]]; then
-        local consulRecursors=$(docker run -it --rm \
-            --net=host \
-            alpine \
-            sed -n "/nameserver/ {s/^.*nameserver[^0-9]*/-recursor /;H}; $ {x;s/\n/ /gp}" /etc/resolv.conf \
-             | sed "s/\r//g"
-        )
-        if [[ "$consulRecursors" ]]; then
-            debug "Consul recursors found on host: $consulRecursors"
-            DOCKER_CONSUL_OPTIONS="$DOCKER_CONSUL_OPTIONS $consulRecursors"
-        fi
+        DOCKER_CONSUL_OPTIONS="$DOCKER_CONSUL_OPTIONS $(consul-recursors <(docker run -it --rm --net=host alpine cat /etc/resolv.conf) $(bridge-ip))"
     fi
     debug "DOCKER_CONSUL_OPTIONS=$DOCKER_CONSUL_OPTIONS"
 }